fix: 코드리뷰 반영 — SHOP 상태 가드/rating 타입/applicationId UK/imageUrls 제약 (STORY-10, KAN-107)
CRITICAL: updateMyShop ACTIVE 상태 가드 추가 (SHOP_007), 죽은 소유권 체크 제거
HIGH: imageUrls @Size(max=2000) 추가 — 전체 JSON 문자열 합산 길이 제한
HIGH: rating float → double (소수점 정밀도 손실 방지)
MEDIUM: application_id UniqueConstraint 추가 — 동시 승인 race condition 차단
MEDIUM: getMyShop SUSPENDED/CLOSED 허용 주석 명시
MEDIUM: 테스트 응답 필드 검증 추가, SUSPENDED 상태 가드 케이스 추가
shopName/category @Size(min=1) 추가 — 빈 문자열 저장 방지

+++ src/main/java/com/chunbaetour/domain/shop/dto/response/ShopResponse.java
@@ -19,7 +19,7 @@ public record ShopResponse(
         String operatingHours,
         String closedDays,
         boolean isCertified,
-        float rating,
+        double rating,
         int reviewCount,
         ShopStatus status
 ) {
